Genes work for 5/7
Issues with Combat powder Animations-Resloved

Made Touch script better looking and more intuitive,
Made it so it tracks whos turn it is and lets the object know its there turn.

Next for me tomorrow is getting the enemys to attack and add Tension to the Tension bar.

(Note for Gene make the Tension variable a float instead of an Integer)